+++ src/PartialAndUndefined.ts
@@ -1,0 +1,1 @@
+export type PartialAndUndefined<T> = { [P in keyof T]?: T[P] | undefined };